Refactor position of Gitlab request parser/validator
